Limit group image editing to users with update permission
This commit restricts the ability to change the group image in the `group_setting_screen.dart` file.

Users without the `group.update` permission will no longer be able to tap the group image to select a new one. The "Tap to select group image" text is also hidden for these users.