Refactor ViewModels to use Group object and update UI
This commit refactors several ViewModels to accept a `Group` object in their constructors instead of a `groupId`. This simplifies the ViewModels and makes it easier to access group information.

The following ViewModels were updated:
- `GroupHomeViewModel`
- `CreateDoctorViewModel`
- `GroupLinkViewModel`
- `GroupMemberViewModel`
- `GroupSettingViewModel`
- `CreateHospitalViewModel`
- `MedicalEntitiesViewModel`

The `router_config.dart` file was updated to pass the `Group` object to these ViewModels.
The `home_screen.dart` file was updated to pass the `Group` object when navigating to the `groupHomeScreen`.
The `group_home_screen.dart` file was updated to display a banner when a group is archived.
The `medical_entities_screen.dart` and `group_setting_screen.dart` files were updated to use the `Group` object from their respective ViewModels.