Add long press gesture to view group image in GroupSettingScreen
This commit introduces a long press gesture to the group image in `group_setting_screen.dart`.

Users can now long press the group image to view it in a dialog with an `InteractiveViewer` for zooming and panning.
The "Long press to view group image" text is also added below the "Tap to select group image" text.